Move Berkeley JE specific classes to a new package
