+++ crawler4j/src/main/java/edu/uci/ics/crawler4j/frontier/je/DocIDServer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package edu.uci.ics.crawler4j.frontier;
+package edu.uci.ics.crawler4j.frontier.je;
 
 import java.util.function.Consumer;
